+++ src/components/orchestrator/Orchestrator.test.tsx
@@ -1,17 +1,29 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
+import { useSearch } from '@tanstack/react-router'
 import { act, waitFor } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import { expect } from 'vitest'
 
 import { MODE_TYPE } from '@/constants/mode'
+import { PAGE_TYPE } from '@/constants/page'
 import { TELEMETRY_EVENT_TYPE } from '@/constants/telemetry'
 import { TelemetryContext } from '@/contexts/TelemetryContext'
+import type { Interrogation } from '@/models/interrogation'
 import { renderWithRouter } from '@/utils/tests'
 
 import { Orchestrator } from './Orchestrator'
 
+vi.mock('@tanstack/react-router', async (importOriginal) => {
+  const actual =
+    (await importOriginal()) as typeof import('@tanstack/react-router')
+  return {
+    ...actual,
+    useSearch: vi.fn(),
+  }
+})
+
 describe('Orchestrator', () => {
-  const interrogation = {
+  const defaultInterrogation = {
     stateData: undefined,
     data: undefined,
     personalization: undefined,
@@ -50,27 +62,33 @@ describe('Orchestrator', () => {
     maxPage: '2',
   }
   const queryClient = new QueryClient()
+
+  const mockGetDepositProof = vi.fn()
+
   const OrchestratorTestWrapper = ({
     mode,
+    isDownloadEnabled = false,
+    initialInterrogation = defaultInterrogation,
   }: {
     mode: MODE_TYPE.COLLECT | MODE_TYPE.REVIEW | MODE_TYPE.VISUALIZE
+    isDownloadEnabled?: boolean
+    initialInterrogation?: Interrogation
   }) => (
     <QueryClientProvider client={queryClient}>
       <Orchestrator
         metadata={metadata}
         mode={mode}
-        initialInterrogation={interrogation}
+        isDownloadEnabled={isDownloadEnabled}
+        initialInterrogation={initialInterrogation}
         // @ts-expect-error: we should have a better lunatic mock
         source={source}
         getReferentiel={() => {
           return new Promise(() => [])
         }}
         updateDataAndStateData={() => {
-          return new Promise<void>(() => {})
+          return new Promise<void>(() => { })
         }}
-        getDepositProof={() => {
-          return new Promise<void>(() => {})
-        }}
+        getDepositProof={mockGetDepositProof}
       />
     </QueryClientProvider>
   )
@@ -81,8 +99,8 @@ describe('Orchestrator', () => {
     renderWithRouter(
       <TelemetryContext.Provider
         value={{
-          isTelemetryDisabled: false,
-          pushEvent: () => {},
+          isTelemetryEnabled: true,
+          pushEvent: () => { },
           setDefaultValues,
         }}
       >
@@ -104,9 +122,9 @@ describe('Orchestrator', () => {
     renderWithRouter(
       <TelemetryContext.Provider
         value={{
-          isTelemetryDisabled: false,
+          isTelemetryEnabled: true,
           pushEvent,
-          setDefaultValues: () => {},
+          setDefaultValues: () => { },
         }}
       >
         <OrchestratorTestWrapper mode={MODE_TYPE.COLLECT} />
@@ -129,9 +147,9 @@ describe('Orchestrator', () => {
     const { getByText } = renderWithRouter(
       <TelemetryContext.Provider
         value={{
-          isTelemetryDisabled: false,
+          isTelemetryEnabled: true,
           pushEvent,
-          setDefaultValues: () => {},
+          setDefaultValues: () => { },
         }}
       >
         <OrchestratorTestWrapper mode={MODE_TYPE.VISUALIZE} />
@@ -149,9 +167,9 @@ describe('Orchestrator', () => {
     const { getByText } = renderWithRouter(
       <TelemetryContext.Provider
         value={{
-          isTelemetryDisabled: false,
+          isTelemetryEnabled: true,
           pushEvent,
-          setDefaultValues: () => {},
+          setDefaultValues: () => { },
         }}
       >
         <OrchestratorTestWrapper mode={MODE_TYPE.REVIEW} />
@@ -169,9 +187,9 @@ describe('Orchestrator', () => {
     const { getByText } = renderWithRouter(
       <TelemetryContext.Provider
         value={{
-          isTelemetryDisabled: true,
+          isTelemetryEnabled: false,
           pushEvent,
-          setDefaultValues: () => {},
+          setDefaultValues: () => { },
         }}
       >
         <OrchestratorTestWrapper mode={MODE_TYPE.COLLECT} />
@@ -189,9 +207,9 @@ describe('Orchestrator', () => {
     const { getByText } = renderWithRouter(
       <TelemetryContext.Provider
         value={{
-          isTelemetryDisabled: false,
+          isTelemetryEnabled: true,
           pushEvent,
-          setDefaultValues: () => {},
+          setDefaultValues: () => { },
         }}
       >
         <OrchestratorTestWrapper mode={MODE_TYPE.COLLECT} />
@@ -215,9 +233,9 @@ describe('Orchestrator', () => {
     const { getByText } = renderWithRouter(
       <TelemetryContext.Provider
         value={{
-          isTelemetryDisabled: false,
+          isTelemetryEnabled: true,
           pushEvent,
-          setDefaultValues: () => {},
+          setDefaultValues: () => { },
         }}
       >
         <OrchestratorTestWrapper mode={MODE_TYPE.COLLECT} />
@@ -258,5 +276,83 @@ describe('Orchestrator', () => {
 
     act(() => getByText('Continue').click())
     expect(document.title).not.toContain('*')
+  })
+
+  it('shows download button when enabled', async () => {
+    const user = userEvent.setup()
+    const { getByText } = renderWithRouter(
+      <OrchestratorTestWrapper
+        mode={MODE_TYPE.COLLECT}
+        isDownloadEnabled={true}
+      />,
+    )
+
+    act(() => getByText('Start').click())
+    act(() => getByText('Continue').click())
+
+    const e = getByText('my-question')
+
+    await user.click(e)
+    await user.keyboard('f')
+
+    act(() => getByText('Continue').click())
+
+    expect(getByText('Download data')).toBeInTheDocument()
+  })
+
+  it('always shows download button in visualize mode', async () => {
+    const user = userEvent.setup()
+    const { getByText } = renderWithRouter(
+      <OrchestratorTestWrapper mode={MODE_TYPE.VISUALIZE} />,
+    )
+
+    act(() => getByText('Start').click())
+    act(() => getByText('Continue').click())
+
+    const e = getByText('my-question')
+
+    await user.click(e)
+    await user.keyboard('f')
+
+    act(() => getByText('Continue').click())
+
+    expect(getByText('Download data')).toBeInTheDocument()
+  })
+
+  it('hides download button when disabled', async () => {
+    const { getByText, queryByText } = renderWithRouter(
+      <OrchestratorTestWrapper mode={MODE_TYPE.COLLECT} />,
+    )
+    act(() => getByText('Start').click())
+    expect(queryByText('Download data')).toBeNull()
+  })
+
+  it('calls getDepositProof with right params', async () => {
+    // Set initial interrogation as validated on end page to directly enable download receipt
+    const initialInterrogation: Interrogation = {
+      ...defaultInterrogation,
+      stateData: { state: 'VALIDATED', currentPage: PAGE_TYPE.END, date: 1 },
+    }
+
+    // Provide encoded param via URL
+    vi.mocked(useSearch).mockReturnValue({
+      surveyUnitCompositeName: 'RnJhbsOnYWlz',
+    })
+
+    const { getByText } = renderWithRouter(
+      <OrchestratorTestWrapper
+        mode={MODE_TYPE.COLLECT}
+        initialInterrogation={initialInterrogation}
+      />,
+    )
+
+    vi.mocked(mockGetDepositProof).mockResolvedValue(undefined)
+
+    // download deposit proof
+    act(() => getByText('Download the receipt').click())
+
+    expect(mockGetDepositProof).toHaveBeenCalledWith({
+      surveyUnitCompositeName: 'RnJhbsOnYWlz',
+    })
   })
 })