+++ src/components/orchestrator/Orchestrator.tsx
@@ -9,7 +9,7 @@ import {
   getArticulationState,
   useLunatic,
 } from '@inseefr/lunatic'
-import { useNavigate } from '@tanstack/react-router'
+import { useNavigate, useSearch } from '@tanstack/react-router'
 import { assert } from 'tsafe/assert'
 
 import { MODE_TYPE } from '@/constants/mode'
@@ -18,6 +18,7 @@ import { useTelemetry } from '@/contexts/TelemetryContext'
 import { useAddPreLogoutAction } from '@/hooks/prelogout'
 import { useBeforeUnload } from '@/hooks/useBeforeUnload'
 import { usePrevious } from '@/hooks/usePrevious'
+import type { GenerateDepositProofParams } from '@/models/api'
 import type { Interrogation } from '@/models/interrogation'
 import type { InterrogationData } from '@/models/interrogationData'
 import type {
@@ -76,6 +77,8 @@ export namespace OrchestratorProps {
     getReferentiel: LunaticGetReferentiel
     /** Interrogation metadata */
     metadata: Metadata
+    /** Indicates if the download button should be enabled */
+    isDownloadEnabled: boolean
   }
 
   export type Visualize = {
@@ -96,14 +99,18 @@ export namespace OrchestratorProps {
       isLogout: boolean
     }) => Promise<void>
     /** Allows user to download a deposit proof PDF */
-    getDepositProof: () => Promise<void>
+    getDepositProof: (params?: GenerateDepositProofParams) => Promise<void>
   }
 }
 
 export function Orchestrator(props: OrchestratorProps) {
-  const { source, getReferentiel, mode, metadata } = props
+  const { source, getReferentiel, mode, isDownloadEnabled, metadata } = props
 
   const navigate = useNavigate()
+
+  const search = useSearch({ strict: false })
+
+  const encodedSurveyUnitCompositeName = search?.surveyUnitCompositeName
 
   const initialInterrogation = computeInterrogation(props.initialInterrogation)
 
@@ -115,7 +122,7 @@ export function Orchestrator(props: OrchestratorProps) {
   const [isTelemetryInitialized, setIsTelemetryInitialized] =
     useState<boolean>(false)
   const {
-    isTelemetryDisabled,
+    isTelemetryEnabled,
     pushEvent,
     setDefaultValues,
     triggerBatchTelemetryCallback,
@@ -345,11 +352,11 @@ export function Orchestrator(props: OrchestratorProps) {
 
   // Telemetry initialization
   useEffect(() => {
-    if (!isTelemetryDisabled && mode === MODE_TYPE.COLLECT) {
+    if (isTelemetryEnabled && mode === MODE_TYPE.COLLECT) {
       setDefaultValues({ idInterrogation: initialInterrogation?.id })
       setIsTelemetryInitialized(true)
     }
-  }, [isTelemetryDisabled, mode, setDefaultValues, initialInterrogation?.id])
+  }, [isTelemetryEnabled, mode, setDefaultValues, initialInterrogation?.id])
 
   // Initialization
   useEffect(() => {
@@ -440,7 +447,9 @@ export function Orchestrator(props: OrchestratorProps) {
       return
     }
     if (mode === MODE_TYPE.COLLECT) {
-      props.getDepositProof()
+      props.getDepositProof({
+        surveyUnitCompositeName: encodedSurveyUnitCompositeName,
+      })
       return
     }
   }
@@ -461,6 +470,7 @@ export function Orchestrator(props: OrchestratorProps) {
           pagination={pagination}
           overview={overview}
           hasArticulation={hasArticulation}
+          isDownloadEnabled={isDownloadEnabled}
           isDirtyState={isDirtyState}
           isSequencePage={isSequencePage(components)}
           bottomContent={